@@ -1,7 +1,7 @@
-import UIKit
 import Flutter
+import UIKit
 
-@UIApplicationMain
+@main
 @objc class AppDelegate: FlutterAppDelegate {
   override func application(
     _ application: UIApplication,
