@@ -1,1 +1,11 @@
-export 'networks/eth.dart';
+import 'dart:js_interop';
+import 'package:blockchain_utils/uuid/uuid.dart';
+import '../models/models.dart';
+import '../utils/utils.dart';
+import 'dart:async';
+import 'package:mrt_native_support/web/mrt_native_web.dart';
+
+part 'core/controller.dart';
+part 'networks/eth.dart';
+part 'networks/tron.dart';
+part 'controller/controller.dart';
