@@ -7,7 +7,13 @@ import 'package:mrt_wallet/wallet/web3/core/permission/models/authenticated.dart
 import 'js_wallet/js_wallet.dart';
 import 'package:mrt_native_support/web/mrt_native_web.dart';
 
+@JS("#OnBackgroundListener_")
+external set OnContentListener(JSFunction? f);
+
 void main(List<String> args) async {
+  if (mrtNull == null) {
+    mrt = MRTWallet(JSObject());
+  }
   final applicationId =
       Web3APPAuthentication.toApplicationId(jsWindow.location.origin);
   if (applicationId == null) {
@@ -16,24 +22,27 @@ void main(List<String> args) async {
 
   final completer = Completer<JSWebviewWallet>();
   bool onActivation(JSWalletEvent data) {
-    final String clientId = mrt.scriptId();
-
-    final walletEvent = data.toEvent();
-    if (walletEvent == null || walletEvent.clientId != clientId) return false;
-    if (walletEvent.type == WalletEventTypes.exception) {
-      final message = Web3ExceptionMessage.deserialize(bytes: walletEvent.data);
-      completer.completeError(message.toWalletError());
-      return false;
-    }
-    if (walletEvent.type != WalletEventTypes.activation) {
-      return false;
-    }
-    final wallet = JSWebviewWallet.initialize(
-        request: walletEvent, clientId: mrt.scriptId());
-    completer.complete(wallet);
+    try {
+      final String clientId = mrt.scriptId;
+      final walletEvent = data.toEvent();
+      if (walletEvent == null || walletEvent.clientId != clientId) return false;
+      if (walletEvent.type == WalletEventTypes.exception) {
+        final message =
+            Web3ExceptionMessage.deserialize(bytes: walletEvent.data);
+        completer.completeError(message.toWalletError());
+        return false;
+      }
+      if (walletEvent.type != WalletEventTypes.activation) {
+        return false;
+      }
+      final wallet =
+          JSWebviewWallet.initialize(request: walletEvent, clientId: clientId);
+      completer.complete(wallet);
+    } catch (e, s) {}
     return true;
   }
 
   mrt.onMrtMessage = onActivation.toJS;
+
   await completer.future;
 }
